Compiles in clang 19 and gcc 14 with -Wall -Wextra*
* For some reason gcc is showing a bunch of warnings about qvariant cast index out of bounds when built in release mode but not in debug (where I was testing before). Will have to investigate that further...

clang 20 is not at all happy with spdlog (not just warnings but errors about constexpr stuff), so that's why I'm stuck on 19 until I'm able to either update or remove spdlog.

Now as far as how it actually works after it's built:

It does seem to run ok for the most part. There are some minor graphics issues, and a lot of GStreamer runtime warnings, and the song store has mysteriously become inoperable. But dragging something from Qt 5.15 all the way to 6.9, and GStreamer 1.16 to 1.24, and through 4 years of compiler updates is bound to cause some issues.